requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cobble==0.1.4
colorama==0.4.6
cryptography==46.0.4
cssselect2==0.8.0
distlib==0.4.0
distro==1.9.0
EbookLib==0.18
einops==0.8.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
filelock==3.20.3
filetype==1.2.0
Flask==3.1.3
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
google-api-core==2.29.0
google-api-python-client==2.112.0
google-auth==2.48.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai==1.61.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
icalendar==6.3.2
identify==2.6.16
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
lxml==6.0.2
mammoth==1.11.0
markdown2==2.5.4
markdownify==1.2.2
marker-pdf==1.10.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.10.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==26.0
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.5.1
playwright==1.58.0
pre_commit==4.5.1
proto-plus==1.27.0
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.1
PyMuPDF==1.27.1
pyparsing==3.3.2
pypdfium2==4.30.0
pyperclip==1.11.0
pyphen==0.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.2.1
python-pptx==1.0.2
PyYAML==6.0.1
RapidFuzz==3.14.3
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
surya-ocr==0.17.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.2
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.36.1
watchdog==3.0.0
wcwidth==0.5.3
weasyprint==63.1
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.6
xlsxwriter==3.2.9
zopfli==0.4.0
